requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
torch==2.3.0
torchvision==0.15.2
pillow==9.5.0
python-multipart==0.0.9